test(cypress): rebalance parallel suites with v5 support
