+++ integration-tests/cypress-esm-config.mjs
@@ -28,8 +28,9 @@ async function runCypress () {
     },
   })
 
-  if (results.totalFailed !== 0) {
-    process.exit(1)
+  const failures = results.totalFailed ?? results.failures ?? 0
+  if (failures !== 0) {
+    process.exit(failures)
   }
 }
