+++ integration-tests/cypress/e2e/flaky-test-retries.js
@@ -1,10 +1,12 @@
 /* eslint-disable */
 describe('flaky test retry', () => {
   let numAttempt = 0
+  const passAttempt = Number(Cypress.env('FLAKY_PASS_ATTEMPT') || 2)
+
   it('eventually passes', () => {
     cy.visit('/')
       .get('.hello-world')
-      .should('have.text', numAttempt++ === 2 ? 'Hello World' : 'Hello Warld')
+      .should('have.text', numAttempt++ === passAttempt ? 'Hello World' : 'Hello Warld')
   })
   it('never passes', () => {
     cy.visit('/')
