+++ integration-tests/cypress/e2e/flaky-with-hooks.cy.js
@@ -1,5 +1,6 @@
 /* eslint-disable */
 let numAttempt = 0
+const passAttempt = Number(Cypress.env('FLAKY_PASS_ATTEMPT') || 2)
 describe('flaky with hooks', () => {
   beforeEach(() => {
     cy.visit('/')
@@ -8,7 +9,7 @@ describe('flaky with hooks', () => {
     // cleanup
   })
   it('eventually passes', () => {
-    cy.get('.hello-world').should('have.text', numAttempt++ === 2 ? 'Hello World' : 'Hello Warld')
+    cy.get('.hello-world').should('have.text', numAttempt++ === passAttempt ? 'Hello World' : 'Hello Warld')
   })
   it('never passes', () => {
     cy.get('.hello-world').should('have.text', 'Hello Warld')
